pyproject.toml
[project]
name = "cc-discord-notifier"
version = "0.1.0"
description = "Claude Code のリリース情報を Discord へ日本語配信するワークフロー"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "pydantic>=2.9",
    "google-genai>=1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-mock>=3.14",
    "ruff>=0.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/notifier"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.lint.per-file-ignores]
"src/notifier/summarizer.py" = ["E501"]  # Gemini への日本語プロンプト原文は改行できない

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
